+++ disma/src/impls/discord/mod.rs
@@ -1,0 +1,8 @@
+pub mod api;
+mod dtos;
+
+mod http_commander;
+pub use http_commander::*;
+
+mod http_querier;
+pub use http_querier::*;
