--- disma/src/infra/discord/mod.rs
@@ -1,3 +1,0 @@
-pub mod api;
-pub mod client;
-pub mod dtos;
