--- disma/src/test/fixtures/existing/mod.rs
@@ -1,4 +1,0 @@
-pub mod category;
-pub mod channel;
-pub mod guild;
-pub mod role;
