--- disma/src/test/utils/vec.rs
@@ -1,8 +1,0 @@
-#[cfg(test)]
-pub mod tests {
-    pub fn assert_contains_exactly_in_any_order<T: PartialEq<T>>(vec: &[T], expected: &Vec<T>) {
-        for expected_item in expected {
-            assert!(vec.contains(expected_item));
-        }
-    }
-}
