+++ disma/src/tests/fixtures/existing/mod.rs
@@ -1,0 +1,11 @@
+mod category;
+pub use category::*;
+
+mod channel;
+pub use channel::*;
+
+mod guild;
+pub use guild::*;
+
+mod role;
+pub use role::*;
