+++ disma/src/utils/mod.rs
@@ -1,3 +1,2 @@
 pub mod http;
 pub mod misc;
-pub mod option;
